refactor(frontend): delete legacy lib/api.ts and route consumers through shared/api (FR-007 US2.7)
